Move classes to Drupal version scopes namespace
