+++ phpstan-baseline.neon
@@ -101,11 +101,11 @@ parameters:
 			path: src/Drupal8/Rector/Deprecation/EntityManagerRector.php
 
 		-
+			message: "#^Doing instanceof PHPStan\\\\Type\\\\ObjectType is error\\-prone and deprecated\\. Use Type\\:\\:isObject\\(\\) or Type\\:\\:getObjectClassNames\\(\\) instead\\.$#"
+			count: 1
+			path: src/Drupal8/Rector/Deprecation/FunctionalTestDefaultThemePropertyRector.php
+
+		-
 			message: "#^Call to an undefined method PHPStan\\\\Type\\\\Type\\:\\:getValue\\(\\)\\.$#"
 			count: 1
 			path: src/Drupal9/Rector/Deprecation/ExtensionPathRector.php
-
-		-
-			message: "#^Doing instanceof PHPStan\\\\Type\\\\ObjectType is error\\-prone and deprecated\\. Use Type\\:\\:isObject\\(\\) or Type\\:\\:getObjectClassNames\\(\\) instead\\.$#"
-			count: 1
-			path: src/Rector/Class_/FunctionalTestDefaultThemePropertyRector.php
